+++ build.sh
@@ -3,23 +3,25 @@
 set -e
 
 # Set version info
-export BOX_VERSION_BASE="1.2.3"
+export BOX_VERSION_BASE="1.2.4"
 
 # Set versions requested of main components (These will be used in Packer and passed to Ansible downstream)
 # The apt package versions can be determined with "apt-get madison <package name>", use the exact version number!
 export BOX_BASE="ilionx/ubuntu2004"
-export BOX_BASE_VERSION="1.1.6-20220503"
-export MINIKUBE_VERSION="1.25.2"
-export DOCKER_VERSION="5:20.10.14~3-0~ubuntu-focal"
-export KUBECTL_VERSION="1.23.6-00"
-export HELM_VERSION="3.8.2"
+export BOX_BASE_VERSION="1.1.7-20220825"
+export MINIKUBE_VERSION="1.26.1"
+export CRIDOCKERD_VERSION="0.2.5"
+export CRITOOLS_VERSION="1.24.2-00"
+export DOCKER_VERSION="5:20.10.17~3-0~ubuntu-focal"
+export KUBECTL_VERSION="1.25.0-00"
+export HELM_VERSION="3.9.4"
 export KUBETAIL_VERSION="1.6.15"
 export KUBELOGS_VERSION="0.0.1"
 
 # Set versions of supported tools, if they don't match, a warning will be shown on screen
-export VIRTUALBOX_VERSION="6.1.34r150636"
-export PACKER_VERSION="1.8.0"
-export VAGRANT_VERSION="2.2.19"
+export VIRTUALBOX_VERSION="6.1.36r152435"
+export PACKER_VERSION="1.8.3"
+export VAGRANT_VERSION="2.3.0"
 
 # Set the Vagrant cloud user and box name (make sure you have admin permissions to, or are the owner of this repository)
 export VAGRANT_CLOUD_BOX_USER="ilionx"